+++ packages/grpc-js-xds/package.json
@@ -42,7 +42,8 @@
     "yargs": "^15.4.1"
   },
   "dependencies": {
-    "@grpc/proto-loader": "^0.6.0-pre14"
+    "@grpc/proto-loader": "^0.6.0-pre14",
+    "google-auth-library": "^7.0.2"
   },
   "peerDependencies": {
     "@grpc/grpc-js": "~1.2.2"
