+++ packages/grpc-js/package.json
@@ -58,7 +58,6 @@
   },
   "dependencies": {
     "@types/node": "^12.12.47",
-    "google-auth-library": "^6.1.1",
     "semver": "^6.2.0"
   },
   "files": [
