+++ packages/grpc-js/src/experimental.ts
@@ -2,7 +2,6 @@ export { trace } from './logging';
 export { Resolver, ResolverListener, registerResolver } from './resolver';
 export { GrpcUri, uriToString } from './uri-parser';
 export { ServiceConfig } from './service-config';
-export { createGoogleDefaultCredentials } from './channel-credentials';
 export { BackoffTimeout } from './backoff-timeout';
 export { LoadBalancer, LoadBalancingConfig, ChannelControlHelper, registerLoadBalancerType, getFirstUsableConfig, validateLoadBalancingConfig } from './load-balancer';
 export { SubchannelAddress, subchannelAddressToString } from './subchannel';
